fixes for int =\= long
git-svn-id: https://yap.svn.sf.net/svnroot/yap/trunk@782 b08c6af1-5177-4d33-ba66-4b1c6b8b522a

+++ C/agc.c
@@ -33,7 +33,7 @@ STATIC_PROTO(void  ConvDBList, (Term, char *,CELL));
 
 static int agc_calls;
 
-static Int agc_collected;
+static unsigned long int agc_collected;
 
 static Int tot_agc_time = 0; /* total time spent in GC */
